+++ src/code/__osMalloc.c
@@ -117,8 +117,8 @@ void __osMallocAddBlock(Arena* arena, void* start, s32 size) {
     ArenaNode* lastNode;
 
     if (start != NULL) {
-        firstNode = (ArenaNode*)ALIGN16((u32)start);
-        diff = (s32)firstNode - (s32)start;
+        firstNode = (ArenaNode*)ALIGN16((uintptr_t)start);
+        diff = (uintptr_t)firstNode - (uintptr_t)start;
         size2 = (size - diff) & ~0xF;
 
         if (size2 > (s32)sizeof(ArenaNode)) {
@@ -174,8 +174,8 @@ void __osMalloc_FreeBlockTest(Arena* arena, ArenaNode* node) {
     u32* iter;
 
     if (__osMalloc_FreeBlockTest_Enable) {
-        start = (u32*)((u32)node + sizeof(ArenaNode));
-        end = (u32*)((u32)start + node2->size);
+        start = (u32*)((uintptr_t)node + sizeof(ArenaNode));
+        end = (u32*)((uintptr_t)start + node2->size);
         iter = start;
 
         while (iter < end) {
@@ -209,7 +209,7 @@ void* __osMalloc_NoLockDebug(Arena* arena, u32 size, const char* file, s32 line)
             }
 
             if (blockSize < iter->size) {
-                newNode = (ArenaNode*)((u32)iter + blockSize);
+                newNode = (ArenaNode*)((uintptr_t)iter + blockSize);
                 newNode->next = ArenaImpl_GetNextBlock(iter);
                 newNode->prev = iter;
                 newNode->size = iter->size - blockSize;
@@ -226,7 +226,7 @@ void* __osMalloc_NoLockDebug(Arena* arena, u32 size, const char* file, s32 line)
 
             iter->isFree = false;
             ArenaImpl_SetDebugInfo(iter, file, line, arena);
-            alloc = (void*)((u32)iter + sizeof(ArenaNode));
+            alloc = (void*)((uintptr_t)iter + sizeof(ArenaNode));
             if (arena->flag & FILL_ALLOCBLOCK) {
                 func_80106860(alloc, BLOCK_ALLOC_MAGIC, size);
             }
@@ -269,7 +269,7 @@ void* __osMallocRDebug(Arena* arena, u32 size, const char* file, s32 line) {
 
             blockSize = ALIGN16(size) + sizeof(ArenaNode);
             if (blockSize < iter->size) {
-                newNode = (ArenaNode*)((u32)iter + (iter->size - size));
+                newNode = (ArenaNode*)((uintptr_t)iter + (iter->size - size));
                 newNode->next = ArenaImpl_GetNextBlock(iter);
                 newNode->prev = iter;
                 newNode->size = size;
@@ -286,7 +286,7 @@ void* __osMallocRDebug(Arena* arena, u32 size, const char* file, s32 line) {
 
             iter->isFree = false;
             ArenaImpl_SetDebugInfo(iter, file, line, arena);
-            allocR = (void*)((u32)iter + sizeof(ArenaNode));
+            allocR = (void*)((uintptr_t)iter + sizeof(ArenaNode));
             if (arena->flag & FILL_ALLOCBLOCK) {
                 func_80106860(allocR, BLOCK_ALLOC_MAGIC, size);
             }
@@ -320,7 +320,7 @@ void* __osMalloc_NoLock(Arena* arena, u32 size) {
             }
 
             if (blockSize < iter->size) {
-                newNode = (ArenaNode*)((u32)iter + blockSize);
+                newNode = (ArenaNode*)((uintptr_t)iter + blockSize);
                 newNode->next = ArenaImpl_GetNextBlock(iter);
                 newNode->prev = iter;
                 newNode->size = iter->size - blockSize;
@@ -337,7 +337,7 @@ void* __osMalloc_NoLock(Arena* arena, u32 size) {
 
             iter->isFree = false;
             ArenaImpl_SetDebugInfo(iter, NULL, 0, arena);
-            alloc = (void*)((u32)iter + sizeof(ArenaNode));
+            alloc = (void*)((uintptr_t)iter + sizeof(ArenaNode));
             if (arena->flag & FILL_ALLOCBLOCK) {
                 func_80106860(alloc, BLOCK_ALLOC_MAGIC, size);
             }
@@ -379,7 +379,7 @@ void* __osMallocR(Arena* arena, u32 size) {
 
             blockSize = ALIGN16(size) + sizeof(ArenaNode);
             if (blockSize < iter->size) {
-                newNode = (ArenaNode*)((u32)iter + (iter->size - size));
+                newNode = (ArenaNode*)((uintptr_t)iter + (iter->size - size));
                 newNode->next = ArenaImpl_GetNextBlock(iter);
                 newNode->prev = iter;
                 newNode->size = size;
@@ -396,7 +396,7 @@ void* __osMallocR(Arena* arena, u32 size) {
 
             iter->isFree = false;
             ArenaImpl_SetDebugInfo(iter, NULL, 0, arena);
-            alloc = (void*)((u32)iter + sizeof(ArenaNode));
+            alloc = (void*)((uintptr_t)iter + sizeof(ArenaNode));
             if (arena->flag & FILL_ALLOCBLOCK) {
                 func_80106860(alloc, BLOCK_ALLOC_MAGIC, size);
             }
@@ -419,7 +419,7 @@ void __osFree_NoLock(Arena* arena, void* ptr) {
         return;
     }
 
-    node = (ArenaNode*)((u32)ptr - sizeof(ArenaNode));
+    node = (ArenaNode*)((uintptr_t)ptr - sizeof(ArenaNode));
     if (node == NULL || node->magic != NODE_MAGIC) {
         // "__osFree: Unauthorized release (%08x)"
         osSyncPrintf(VT_COL(RED, WHITE) "__osFree:不正解放(%08x)\n" VT_RST, ptr);
@@ -442,11 +442,11 @@ void __osFree_NoLock(Arena* arena, void* ptr) {
     ArenaImpl_SetDebugInfo(node, NULL, 0, arena);
 
     if (arena->flag & FILL_FREEBLOCK) {
-        func_80106860((u32)node + sizeof(ArenaNode), BLOCK_FREE_MAGIC, node->size);
+        func_80106860((uintptr_t)node + sizeof(ArenaNode), BLOCK_FREE_MAGIC, node->size);
     }
 
     newNext = next;
-    if ((u32)next == (u32)node + sizeof(ArenaNode) + node->size && next->isFree) {
+    if ((uintptr_t)next == (uintptr_t)node + sizeof(ArenaNode) + node->size && next->isFree) {
         newNext = ArenaImpl_GetNextBlock(next);
         if (newNext != NULL) {
             newNext->prev = node;
@@ -460,7 +460,7 @@ void __osFree_NoLock(Arena* arena, void* ptr) {
         next = newNext;
     }
 
-    if (prev != NULL && prev->isFree && (u32)node == (u32)prev + sizeof(ArenaNode) + prev->size) {
+    if (prev != NULL && prev->isFree && (uintptr_t)node == (uintptr_t)prev + sizeof(ArenaNode) + prev->size) {
         if (next) {
             next->prev = prev;
         }
@@ -488,7 +488,7 @@ void __osFree_NoLockDebug(Arena* arena, void* ptr, const char* file, s32 line) {
         return;
     }
 
-    node = (ArenaNode*)((u32)ptr - sizeof(ArenaNode));
+    node = (ArenaNode*)((uintptr_t)ptr - sizeof(ArenaNode));
     if (node == NULL || node->magic != NODE_MAGIC) {
         // "__osFree: Unauthorized release (%08x)"
         osSyncPrintf(VT_COL(RED, WHITE) "__osFree:不正解放(%08x) [%s:%d ]\n" VT_RST, ptr, file, line);
@@ -512,11 +512,11 @@ void __osFree_NoLockDebug(Arena* arena, void* ptr, const char* file, s32 line) {
     ArenaImpl_SetDebugInfo(node, file, line, arena);
 
     if (arena->flag & FILL_FREEBLOCK) {
-        func_80106860((u32)node + sizeof(ArenaNode), BLOCK_FREE_MAGIC, node->size);
+        func_80106860((uintptr_t)node + sizeof(ArenaNode), BLOCK_FREE_MAGIC, node->size);
     }
 
     newNext = node->next;
-    if ((u32)next == (u32)node + sizeof(ArenaNode) + node->size && next->isFree) {
+    if ((uintptr_t)next == (uintptr_t)node + sizeof(ArenaNode) + node->size && next->isFree) {
         newNext = ArenaImpl_GetNextBlock(next);
         if (newNext != NULL) {
             newNext->prev = node;
@@ -530,7 +530,7 @@ void __osFree_NoLockDebug(Arena* arena, void* ptr, const char* file, s32 line) {
         next = newNext;
     }
 
-    if (prev != NULL && prev->isFree && (u32)node == (u32)prev + sizeof(ArenaNode) + prev->size) {
+    if (prev != NULL && prev->isFree && (uintptr_t)node == (uintptr_t)prev + sizeof(ArenaNode) + prev->size) {
         if (next != NULL) {
             next->prev = prev;
         }
@@ -572,19 +572,19 @@ void* __osRealloc(Arena* arena, void* ptr, u32 newSize) {
         __osFree_NoLock(arena, ptr);
         ptr = NULL;
     } else {
-        node = (ArenaNode*)((u32)ptr - sizeof(ArenaNode));
+        node = (ArenaNode*)((uintptr_t)ptr - sizeof(ArenaNode));
         if (newSize == node->size) {
             // "Does nothing because the memory block size does not change"
             osSyncPrintf("メモリブロックサイズが変わらないためなにもしません\n");
         } else if (node->size < newSize) {
             next = ArenaImpl_GetNextBlock(node);
             sizeDiff = newSize - node->size;
-            if ((u32)next == ((u32)node + node->size + sizeof(ArenaNode)) && next->isFree && next->size >= sizeDiff) {
+            if ((uintptr_t)next == ((uintptr_t)node + node->size + sizeof(ArenaNode)) && next->isFree && next->size >= sizeDiff) {
                 // "Merge because there is a free block after the current memory block"
                 osSyncPrintf("現メモリブロックの後ろにフリーブロックがあるので結合します\n");
                 next->size -= sizeDiff;
                 overNext = ArenaImpl_GetNextBlock(next);
-                newNext = (ArenaNode*)((u32)next + sizeDiff);
+                newNext = (ArenaNode*)((uintptr_t)next + sizeDiff);
                 if (overNext != NULL) {
                     overNext->prev = newNext;
                 }
@@ -607,7 +607,7 @@ void* __osRealloc(Arena* arena, void* ptr, u32 newSize) {
                 blockSize = ALIGN16(newSize) + sizeof(ArenaNode);
                 // "Increased free block behind current memory block"
                 osSyncPrintf("現メモリブロックの後ろのフリーブロックを大きくしました\n");
-                newNext2 = (ArenaNode*)((u32)node + blockSize);
+                newNext2 = (ArenaNode*)((uintptr_t)node + blockSize);
                 localCopy = *next2;
                 *newNext2 = localCopy;
                 newNext2->size += node->size - newSize;
@@ -621,7 +621,7 @@ void* __osRealloc(Arena* arena, void* ptr, u32 newSize) {
                 blockSize = ALIGN16(newSize) + sizeof(ArenaNode);
                 // "Generated because there is no free block after the current memory block"
                 osSyncPrintf("現メモリブロックの後ろにフリーブロックがないので生成します\n");
-                newNext2 = (ArenaNode*)((u32)node + blockSize);
+                newNext2 = (ArenaNode*)((uintptr_t)node + blockSize);
                 newNext2->next = ArenaImpl_GetNextBlock(node);
                 newNext2->prev = node;
                 newNext2->size = node->size - blockSize;
@@ -701,7 +701,7 @@ void __osDisplayArena(Arena* arena) {
     while (iter != NULL) {
         if (iter != NULL && iter->magic == NODE_MAGIC) {
             next = iter->next;
-            osSyncPrintf("%08x-%08x%c %s %08x", iter, ((u32)iter + sizeof(ArenaNode) + iter->size),
+            osSyncPrintf("%08x-%08x%c %s %08x", iter, ((uintptr_t)iter + sizeof(ArenaNode) + iter->size),
                          (next == NULL) ? '$' : (iter != next->prev ? '!' : ' '),
                          iter->isFree ? "空き" : "確保", //? "Free" : "Secure"
                          iter->size);
@@ -761,7 +761,7 @@ void ArenaImpl_FaultClient(Arena* arena) {
     while (iter != NULL) {
         if (iter != NULL && iter->magic == NODE_MAGIC) {
             next = iter->next;
-            FaultDrawer_Printf("%08x-%08x%c %s %08x", iter, ((u32)iter + sizeof(ArenaNode) + iter->size),
+            FaultDrawer_Printf("%08x-%08x%c %s %08x", iter, ((uintptr_t)iter + sizeof(ArenaNode) + iter->size),
                                (!next) ? '$' : (iter != next->prev ? '!' : ' '), iter->isFree ? "F" : "A", iter->size);
 
             FaultDrawer_Printf("\n");
