+++ src/code/code_800FC620.c
@@ -20,9 +20,9 @@ char D_80134488[0x18] = {
     0xFF, 0x80, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0x00, 0x00, 0x00,
 };
 
-s32 Overlay_Load(u32 vRomStart, u32 vRomEnd, void* vRamStart, void* vRamEnd, void* allocatedVRamAddr) {
+s32 Overlay_Load(uintptr_t vRomStart, uintptr_t vRomEnd, void* vRamStart, void* vRamEnd, void* allocatedVRamAddr) {
     s32 pad;
-    u32 end;
+    uintptr_t end;
     u32 bssSize;
     OverlayRelocationSection* ovl;
     u32 relocCnt;
@@ -38,16 +38,16 @@ s32 Overlay_Load(u32 vRomStart, u32 vRomEnd, void* vRamStart, void* vRamEnd, voi
         size = vRomEnd - vRomStart;
         // "DMA transfer of TEXT, DATA, RODATA + rel (%08x-%08x)"
         osSyncPrintf("TEXT,DATA,RODATA+relをＤＭＡ転送します(%08x-%08x)\n", allocatedVRamAddr,
-                     (u32)allocatedVRamAddr + size);
+                     (uintptr_t)allocatedVRamAddr + size);
     }
 
     size = vRomEnd - vRomStart;
-    end = (u32)allocatedVRamAddr + size;
-    DmaMgr_SendRequest0((u32)allocatedVRamAddr, vRomStart, size);
+    end = (uintptr_t)allocatedVRamAddr + size;
+    DmaMgr_SendRequest0((uintptr_t)allocatedVRamAddr, vRomStart, size);
 
     ovlOffset = ((s32*)end)[-1];
 
-    ovl = (OverlayRelocationSection*)((u32)end - ovlOffset);
+    ovl = (OverlayRelocationSection*)((uintptr_t)end - ovlOffset);
     if (gOverlayLogSeverity >= 3) {
         osSyncPrintf("TEXT(%08x), DATA(%08x), RODATA(%08x), BSS(%08x)\n", ovl->textSize, ovl->dataSize, ovl->rodataSize,
                      ovl->bssSize);
@@ -73,15 +73,15 @@ s32 Overlay_Load(u32 vRomStart, u32 vRomEnd, void* vRamStart, void* vRamEnd, voi
         (void)relocCnt; // suppresses set but unused warning
     }
 
-    size = (u32)&ovl->relocations[ovl->nRelocations] - (u32)ovl;
+    size = (uintptr_t)&ovl->relocations[ovl->nRelocations] - (uintptr_t)ovl;
     if (gOverlayLogSeverity >= 3) {
         // "Clear REL area (%08x-%08x)"
-        osSyncPrintf("REL領域をクリアします(%08x-%08x)\n", ovl, (u32)ovl + size);
+        osSyncPrintf("REL領域をクリアします(%08x-%08x)\n", ovl, (uintptr_t)ovl + size);
     }
 
     bzero(ovl, size);
 
-    size = (u32)vRamEnd - (u32)vRamStart;
+    size = (uintptr_t)vRamEnd - (uintptr_t)vRamStart;
     osWritebackDCache(allocatedVRamAddr, size);
     osInvalICache(allocatedVRamAddr, size);
 
@@ -109,50 +109,50 @@ void func_800FC83C(void* ptr) {
 }
 
 void func_800FC868(void* blk, u32 nBlk, u32 blkSize, arg3_800FC868 arg3) {
-    u32 pos;
+    uintptr_t pos;
 
-    for (pos = (u32)blk; pos < (u32)blk + (nBlk * blkSize); pos = (u32)pos + (blkSize & ~0)) {
+    for (pos = (uintptr_t)blk; pos < (uintptr_t)blk + (nBlk * blkSize); pos = (uintptr_t)pos + (blkSize & ~0)) {
         arg3((void*)pos);
     }
 }
 
 void func_800FC8D8(void* blk, u32 nBlk, s32 blkSize, arg3_800FC8D8 arg3) {
-    u32 pos;
+    uintptr_t pos;
 
-    for (pos = (u32)blk; pos < (u32)blk + (nBlk * blkSize); pos = (u32)pos + (blkSize & ~0)) {
+    for (pos = (uintptr_t)blk; pos < (uintptr_t)blk + (nBlk * blkSize); pos = (uintptr_t)pos + (blkSize & ~0)) {
         arg3((void*)pos, 2);
     }
 }
 
 void* func_800FC948(void* blk, u32 nBlk, u32 blkSize, arg3_800FC948 arg3) {
-    u32 pos;
+    uintptr_t pos;
 
     if (blk == NULL) {
         blk = func_800FC800(nBlk * blkSize);
     }
 
     if (blk != NULL && arg3 != NULL) {
-        pos = (u32)blk;
-        while (pos < (u32)blk + (nBlk * blkSize)) {
+        pos = (uintptr_t)blk;
+        while (pos < (uintptr_t)blk + (nBlk * blkSize)) {
             arg3((void*)pos, 0, 0, 0, 0, 0, 0, 0, 0);
-            pos = (u32)pos + (blkSize & ~0);
+            pos = (uintptr_t)pos + (blkSize & ~0);
         }
     }
     return blk;
 }
 
 void func_800FCA18(void* blk, u32 nBlk, u32 blkSize, arg3_800FCA18 arg3, s32 arg4) {
-    u32 pos;
-    u32 end;
+    uintptr_t pos;
+    uintptr_t end;
     s32 masked_arg2;
 
     if (blk == 0) {
         return;
     }
     if (arg3 != 0) {
-        end = (u32)blk;
+        end = (uintptr_t)blk;
         masked_arg2 = (s32)(blkSize & ~0);
-        pos = (u32)end + (nBlk * blkSize);
+        pos = (uintptr_t)end + (nBlk * blkSize);
 
         if (masked_arg2) {}
